# Conflicts:
#	typo3/sysext/about/composer.json
#	typo3/sysext/about/ext_emconf.php
#	typo3/sysext/adminpanel/composer.json
#	typo3/sysext/adminpanel/ext_emconf.php
#	typo3/sysext/backend/Tests/Functional/Configuration/SiteConfiguration/Fixtures/Extensions/conf_overriding/a/ext_emconf.php
#	typo3/sysext/backend/Tests/Functional/Configuration/SiteConfiguration/Fixtures/Extensions/conf_overriding/b/ext_emconf.php
#	typo3/sysext/backend/composer.json
#	typo3/sysext/backend/ext_emconf.php
#	typo3/sysext/belog/composer.json
#	typo3/sysext/belog/ext_emconf.php
#	typo3/sysext/beuser/composer.json
#	typo3/sysext/beuser/ext_emconf.php
#	typo3/sysext/core/Classes/Core/SystemEnvironmentBuilder.php
#	typo3/sysext/core/Classes/Database/QueryView.php
#	typo3/sysext/core/Classes/Exception/Archive/ExtractException.php
#	typo3/sysext/core/Classes/Service/Archive/ZipService.php
#	typo3/sysext/core/Tests/Functional/Category/Collection/Fixtures/Extensions/test/ext_emconf.php
#	typo3/sysext/core/Tests/Functional/Database/Fixtures/Extensions/test_expressionbuilder/ext_emconf.php
#	typo3/sysext/core/Tests/Functional/Database/QueryGeneratorTest.php
#	typo3/sysext/core/Tests/Functional/Fixtures/Extensions/irre_tutorial/ext_emconf.php
#	typo3/sysext/core/Tests/Functional/Fixtures/Extensions/test_datahandler/ext_emconf.php
#	typo3/sysext/core/Tests/Functional/Fixtures/Extensions/test_meta/ext_emconf.php
#	typo3/sysext/core/Tests/Functional/Fixtures/Extensions/test_resources/ext_emconf.php
#	typo3/sysext/core/Tests/Functional/Service/Archive/ZipServiceTest.php
#	typo3/sysext/core/Tests/Unit/Utility/Fixtures/ext_emconf.php
#	typo3/sysext/core/composer.json
#	typo3/sysext/core/ext_emconf.php
#	typo3/sysext/extbase/Tests/Functional/Fixtures/Extensions/blog_example/ext_emconf.php
#	typo3/sysext/extbase/Tests/Functional/Fixtures/Extensions/class_overriding/a/ext_emconf.php
#	typo3/sysext/extbase/Tests/Functional/Fixtures/Extensions/class_overriding/b/ext_emconf.php
#	typo3/sysext/extbase/composer.json
#	typo3/sysext/extbase/ext_emconf.php
#	typo3/sysext/extensionmanager/Classes/Utility/FileHandlingUtility.php
#	typo3/sysext/extensionmanager/composer.json
#	typo3/sysext/extensionmanager/ext_emconf.php
#	typo3/sysext/felogin/Classes/Controller/FrontendLoginController.php
#	typo3/sysext/felogin/composer.json
#	typo3/sysext/felogin/ext_emconf.php
#	typo3/sysext/filelist/composer.json
#	typo3/sysext/filelist/ext_emconf.php
#	typo3/sysext/filemetadata/composer.json
#	typo3/sysext/filemetadata/ext_emconf.php
#	typo3/sysext/fluid/Tests/Functional/Fixtures/Extensions/fluid_test/ext_emconf.php
#	typo3/sysext/fluid/composer.json
#	typo3/sysext/fluid/ext_emconf.php
#	typo3/sysext/fluid_styled_content/composer.json
#	typo3/sysext/fluid_styled_content/ext_emconf.php
#	typo3/sysext/form/Tests/Functional/Hooks/Fixtures/test_resources/ext_emconf.php
#	typo3/sysext/form/composer.json
#	typo3/sysext/form/ext_emconf.php
#	typo3/sysext/frontend/Classes/ContentObject/ContentObjectRenderer.php
#	typo3/sysext/frontend/Tests/Functional/SiteHandling/Fixtures/TypoLinkScenario.yaml
#	typo3/sysext/frontend/Tests/Functional/SiteHandling/TypoLinkTest.php
#	typo3/sysext/frontend/composer.json
#	typo3/sysext/frontend/ext_emconf.php
#	typo3/sysext/impexp/Tests/Functional/Fixtures/Extensions/template_extension/ext_emconf.php
#	typo3/sysext/impexp/composer.json
#	typo3/sysext/impexp/ext_emconf.php
#	typo3/sysext/indexed_search/composer.json
#	typo3/sysext/indexed_search/ext_emconf.php
#	typo3/sysext/info/composer.json
#	typo3/sysext/info/ext_emconf.php
#	typo3/sysext/install/composer.json
#	typo3/sysext/install/ext_emconf.php
#	typo3/sysext/linkvalidator/composer.json
#	typo3/sysext/linkvalidator/ext_emconf.php
#	typo3/sysext/lowlevel/composer.json
#	typo3/sysext/lowlevel/ext_emconf.php
#	typo3/sysext/opendocs/composer.json
#	typo3/sysext/opendocs/ext_emconf.php
#	typo3/sysext/recordlist/composer.json
#	typo3/sysext/recordlist/ext_emconf.php
#	typo3/sysext/recycler/composer.json
#	typo3/sysext/recycler/ext_emconf.php
#	typo3/sysext/redirects/composer.json
#	typo3/sysext/redirects/ext_emconf.php
#	typo3/sysext/reports/composer.json
#	typo3/sysext/reports/ext_emconf.php
#	typo3/sysext/rte_ckeditor/composer.json
#	typo3/sysext/rte_ckeditor/ext_emconf.php
#	typo3/sysext/scheduler/composer.json
#	typo3/sysext/scheduler/ext_emconf.php
#	typo3/sysext/seo/composer.json
#	typo3/sysext/seo/ext_emconf.php
#	typo3/sysext/setup/composer.json
#	typo3/sysext/setup/ext_emconf.php
#	typo3/sysext/sys_note/composer.json
#	typo3/sysext/sys_note/ext_emconf.php
#	typo3/sysext/t3editor/composer.json
#	typo3/sysext/t3editor/ext_emconf.php
#	typo3/sysext/tstemplate/composer.json
#	typo3/sysext/tstemplate/ext_emconf.php
#	typo3/sysext/viewpage/composer.json
#	typo3/sysext/viewpage/ext_emconf.php
#	typo3/sysext/workspaces/composer.json
#	typo3/sysext/workspaces/ext_emconf.php

+++ typo3/sysext/install/Classes/Service/LanguagePackService.php
@@ -1,6 +1,6 @@
 <?php
-declare(strict_types = 1);
-namespace TYPO3\CMS\Install\Service;
+
+declare(strict_types=1);
 
 /*
  * This file is part of the TYPO3 CMS project.
@@ -15,10 +15,14 @@ namespace TYPO3\CMS\Install\Service;
  * The TYPO3 project - inspiring people to share!
  */
 
+namespace TYPO3\CMS\Install\Service;
+
 use Psr\EventDispatcher\EventDispatcherInterface;
+use Psr\Log\LoggerAwareInterface;
+use Psr\Log\LoggerAwareTrait;
 use Symfony\Component\Finder\Finder;
-use TYPO3\CMS\Core\Configuration\Features;
 use TYPO3\CMS\Core\Core\Environment;
+use TYPO3\CMS\Core\Http\RequestFactory;
 use TYPO3\CMS\Core\Http\Uri;
 use TYPO3\CMS\Core\Localization\Locales;
 use TYPO3\CMS\Core\Package\PackageManager;
@@ -27,6 +31,7 @@ use TYPO3\CMS\Core\Service\Archive\ZipService;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Utility\PathUtility;
+use TYPO3\CMS\Install\Service\Event\ModifyLanguagePackRemoteBaseUrlEvent;
 
 /**
  * Service class handling language pack details
@@ -34,8 +39,10 @@ use TYPO3\CMS\Core\Utility\PathUtility;
  *
  * @internal This class is only meant to be used within EXT:install and is not part of the TYPO3 Core API.
  */
-class LanguagePackService
+class LanguagePackService implements LoggerAwareInterface
 {
+    use LoggerAwareTrait;
+
     /**
      * @var Locales
      */
@@ -51,14 +58,24 @@ class LanguagePackService
      */
     protected $eventDispatcher;
 
-    private const DEFAULT_LANGUAGE_PACK_URL = 'https://typo3.org/fileadmin/ter/';
-    private const BETA_LANGUAGE_PACK_URL = 'https://beta-translation.typo3.org/fileadmin/ter/';
+    /**
+     * @var RequestFactory
+     */
+    protected $requestFactory;
 
-    public function __construct(EventDispatcherInterface $eventDispatcher = null)
+    private const OLD_LANGUAGE_PACK_URLS = [
+        'https://typo3.org/fileadmin/ter/',
+        'https://beta-translation.typo3.org/fileadmin/ter/',
+        'https://localize.typo3.org/fileadmin/ter/'
+    ];
+    private const LANGUAGE_PACK_URL = 'https://localize.typo3.org/xliff/';
+
+    public function __construct(EventDispatcherInterface $eventDispatcher, RequestFactory $requestFactory)
     {
-        $this->eventDispatcher = $eventDispatcher ?? GeneralUtility::getContainer()->get(EventDispatcherInterface::class);
+        $this->eventDispatcher = $eventDispatcher;
         $this->locales = GeneralUtility::makeInstance(Locales::class);
         $this->registry = GeneralUtility::makeInstance(Registry::class);
+        $this->requestFactory = $requestFactory;
     }
 
     /**
@@ -148,12 +165,19 @@ class LanguagePackService
                 $EM_CONF = [];
                 include $path . 'ext_emconf.php';
                 $title = $EM_CONF[$key]['title'] ?? $title;
+
+                $state = $EM_CONF[$key]['state'] ?? '';
+                if ($state === 'excludeFromUpdates') {
+                    continue;
+                }
             }
             $extension = [
                 'key' => $key,
                 'title' => $title,
-                'icon' => PathUtility::stripPathSitePrefix(ExtensionManagementUtility::getExtensionIcon($path, true)),
             ];
+            if (!empty(ExtensionManagementUtility::getExtensionIcon($path, false))) {
+                $extension['icon'] = PathUtility::stripPathSitePrefix(ExtensionManagementUtility::getExtensionIcon($path, true));
+            }
             $extension['packs'] = [];
             foreach ($activeLanguages as $iso) {
                 $isLanguagePackDownloaded = is_dir(Environment::getLabelsPath() . '/' . $iso . '/' . $key . '/');
@@ -184,19 +208,30 @@ class LanguagePackService
      */
     public function updateMirrorBaseUrl(): string
     {
+        $repositoryUrl = 'https://repositories.typo3.org/mirrors.xml.gz';
         $downloadBaseUrl = false;
         try {
-            $xmlContent = GeneralUtility::getUrl('https://repositories.typo3.org/mirrors.xml.gz');
-            $xmlContent = GeneralUtility::xml2array(@gzdecode($xmlContent));
-            if (!empty($xmlContent['mirror']['host']) && !empty($xmlContent['mirror']['path'])) {
-                $downloadBaseUrl = 'https://' . $xmlContent['mirror']['host'] . $xmlContent['mirror']['path'];
+            $response = $this->requestFactory->request($repositoryUrl);
+            if ($response->getStatusCode() === 200) {
+                $xmlContent = @gzdecode($response->getBody()->getContents());
+                if (!empty($xmlContent['mirror']['host']) && !empty($xmlContent['mirror']['path'])) {
+                    $downloadBaseUrl = 'https://' . $xmlContent['mirror']['host'] . $xmlContent['mirror']['path'];
+                }
+            } else {
+                $this->logger->warning(sprintf(
+                    'Requesting %s was not successful, got status code %d (%s)',
+                    $repositoryUrl,
+                    $response->getStatusCode(),
+                    $response->getReasonPhrase()
+                ));
             }
         } catch (\Exception $e) {
             // Catch generic exception, fallback handled below
+            $this->logger->error('Failed to download list of mirrors', ['exception' => $e]);
         }
         if (empty($downloadBaseUrl)) {
             // Hard coded fallback if something went wrong fetching & parsing mirror list
-            $downloadBaseUrl = self::DEFAULT_LANGUAGE_PACK_URL;
+            $downloadBaseUrl = self::LANGUAGE_PACK_URL;
         }
         $this->registry->set('languagePacks', 'baseUrl', $downloadBaseUrl);
         return $downloadBaseUrl;
@@ -236,23 +271,22 @@ class LanguagePackService
             throw new \RuntimeException('Language pack baseUrl not found', 1520169691);
         }
 
-        if ($languagePackBaseUrl === self::DEFAULT_LANGUAGE_PACK_URL
-            && GeneralUtility::makeInstance(Features::class)->isFeatureEnabled('betaTranslationServer')) {
-            $languagePackBaseUrl = self::BETA_LANGUAGE_PACK_URL;
+        if (in_array($languagePackBaseUrl, self::OLD_LANGUAGE_PACK_URLS, true)) {
+            $languagePackBaseUrl = self::LANGUAGE_PACK_URL;
         }
 
         // Allow to modify the base url on the fly
-        $event = $this->eventDispatcher->dispatch(new Event\ModifyLanguagePackRemoteBaseUrlEvent(new Uri($languagePackBaseUrl), $key));
+        $event = $this->eventDispatcher->dispatch(new ModifyLanguagePackRemoteBaseUrlEvent(new Uri($languagePackBaseUrl), $key));
         $languagePackBaseUrl = $event->getBaseUrl();
         $path = ExtensionManagementUtility::extPath($key);
         $majorVersion = explode('.', TYPO3_branch)[0];
         if (strpos($path, '/sysext/') !== false) {
             // This is a system extension and the package URL should be adapted to have different packs per core major version
-            // https://typo3.org/fileadmin/ter/b/a/backend-l10n/backend-l10n-fr.v9.zip
+            // https://localize.typo3.org/xliff/b/a/backend-l10n/backend-l10n-fr.v9.zip
             $packageUrl = $key[0] . '/' . $key[1] . '/' . $key . '-l10n/' . $key . '-l10n-' . $iso . '.v' . $majorVersion . '.zip';
         } else {
             // Typical non sysext path, Hungarian:
-            // https://typo3.org/fileadmin/ter/a/n/anextension-l10n/anextension-l10n-hu.zip
+            // https://localize.typo3.org/xliff/a/n/anextension-l10n/anextension-l10n-hu.zip
             $packageUrl = $key[0] . '/' . $key[1] . '/' . $key . '-l10n/' . $key . '-l10n-' . $iso . '.zip';
         }
 
@@ -266,20 +300,30 @@ class LanguagePackService
 
         $operationResult = false;
         try {
-            $languagePackContent = GeneralUtility::getUrl($languagePackBaseUrl . $packageUrl);
-            if (!empty($languagePackContent)) {
-                $operationResult = true;
-                if ($packExists) {
-                    $operationResult = GeneralUtility::rmdir($absoluteExtractionPath, true);
+            $response = $this->requestFactory->request($languagePackBaseUrl . $packageUrl);
+            if ($response->getStatusCode() === 200) {
+                $languagePackContent = $response->getBody()->getContents();
+                if (!empty($languagePackContent)) {
+                    $operationResult = true;
+                    if ($packExists) {
+                        $operationResult = GeneralUtility::rmdir($absoluteExtractionPath, true);
+                    }
+                    if ($operationResult) {
+                        GeneralUtility::mkdir_deep(Environment::getVarPath() . '/transient/');
+                        $operationResult = GeneralUtility::writeFileToTypo3tempDir($absolutePathToZipFile, $languagePackContent) === null;
+                    }
+                    $this->unzipTranslationFile($absolutePathToZipFile, $absoluteLanguagePath);
+                    if ($operationResult) {
+                        $operationResult = unlink($absolutePathToZipFile);
+                    }
                 }
-                if ($operationResult) {
-                    GeneralUtility::mkdir_deep(Environment::getVarPath() . '/transient/');
-                    $operationResult = GeneralUtility::writeFileToTypo3tempDir($absolutePathToZipFile, $languagePackContent) === null;
-                }
-                $this->unzipTranslationFile($absolutePathToZipFile, $absoluteLanguagePath);
-                if ($operationResult) {
-                    $operationResult = unlink($absolutePathToZipFile);
-                }
+            } else {
+                $this->logger->warning(sprintf(
+                    'Requesting %s was not successful, got status code %d (%s)',
+                    $languagePackBaseUrl . $packageUrl,
+                    $response->getStatusCode(),
+                    $response->getReasonPhrase()
+                ));
             }
         } catch (\Exception $e) {
             $operationResult = false;
